Implement parsing for flow enums
Progress toward #701

This is based largely off of the Babel implementation, and handles various cases
with ellipses, `of` declarations, assignments, and enums as part of named and
default exports.

I also changed the token test to make it easier to copy and paste the error
output to get a starting point for the tokens.