Implement Flow enums
Fixes #701

There were a few cases to handle around different types of enums and whethery
they are exported or default export and targeting ESM vs CJS, but overall the
code transform behavior is pretty straightforward. One point of complexity was
that the imports transform needed to opt out of handling for the default export
case so that the enum transform can handle it.

Some caveats in the implementation:
* The snippet `require("flow-enums-runtime")` is not currently configurable like
  it is for the babel plugin. It wouldn't be hard to add this as a new option,
  but I think it's best to leave out config options unless they're needed.
* Flow enums have quite a few constraints which are validated by the babel
  parser and which Sucrase does not validate, such as initializers being
  all-or-nothing, initializers always being literals, and initializers agreeing
  with the enum type. Currently, Sucrase doesn't give a very helpful error when
  these assumptions aren't true, e.g. the transform step can overrun the enum
  body and crash unexpectedly later. This hopefully isn't such a big deal since
  Flow will give a helpful error, but there are probably ways to make the
  Sucrase errors more helpful without sacrificing performance.