Add test script to try Sucrase against Babel test cases (#720)
As I was porting Babel changes and reworking bits of the parser, I was worried
that I might introduce regressions in various language corner cases, so it
seemed like a good idea to extend the test suite to more proactively find known
tricky syntax cases. The Babel test suite itself has many fixtures showing off
various syntax cases, so this PR adds a script that clones the Babel repo and
tries Sucrase on every babel-parser test case where the parse is expected to
pass. If Sucrase fails, we try Babel with a Sucrase-like configuration, since
Babel tests somtimes assume sloppy mode or experimental features that Sucrase
doesn't support. For now, the test runner is just its own file rather than using
a real test framework, but we run it in CI to avoid regressions on future code
changes.

This also refactors the directory structure to create a "spec compliance tests"
directory that includes this new one and test262. In the future, it might have
a similar runner for the TypeScript test suite and possibly other test suites.