Refactor jsx-test to better support testing the automatic runtime (#738)
Since the automatic runtime effectively doubles the number of cases needing test
coverage, plus has a number of cases within itself, this PR refactors the JSX
tests to make it easy to declare the result for multiple configurations in the
same test. This should make it smoother to write the automatic case for each
test where it feels relevant, plus makes it easier to test ESM vs CJS and dev vs
prod for the same syntax case.