Add support for automatic JSX runtime (#740)
Fixes #585

Tech plan that includes many major details:
https://github.com/alangpierce/sucrase/wiki/JSX-Automatic-Runtime-Transform-Technical-Plan

This PR adds two new options:
* `jsxRuntime`, which can be "classic" (the existing behavior) or "automatic"
  (the transform behavior released with React 17).
* `jsxImportSource`, which configures the import prefix when using the automatic
  runtime.

This PR also adds docs for the new options and makes them available to demo on the website.

While it may have been possible to split the `JSXTransform` code into multiple files, the
configuration space is intertwined enough that it seemed better for now to just have a unified
implementation for now. This leads to a little extra abstraction when resolving function names
as well as some special cases about where to place commas, but the overall flow of logic is the
same.

Unfortunately, it looks like the new implementation is slower than the old one, regardless of
configuration, which isn't too surprising given the added complexity. Some measurements,
using a higher-scale variation of `yarn benchmark sample` (which specifically tests scale for
the JSX transform):
* Before this PR: 968k lines per second
* Classic runtime after this PR: 925k lines per second (4% slower than before)
* Automatic runtime after this PR: 901k lines per second (7% slower than before)

There's certainly some possible room for improvement, e.g. by more aggressively caching/
precomputing resolved names or duplicating and individually simplifying the code path for
each configuration, but that can be done as future work.