Add integration tests for jest plugin (#751)
This PR reworks and extends the integration test suite to have two tests for the Jest plugin, and hopefully is flexible enough for additional cases in the future. Like the ts-node test suite, it relies on test cases that are defined by directories.

Also update integration tests to no longer use `--delay` and fix some other small issues with integration tests.