Switch Jest plugin to use a peer dependency for Sucrase
The plugin previously had a regular dependency, which is useful in that it makes
installation simpler, but is a bit unusual for these sorts of integration
packages, and makes it harder to manage dependencies when you want to pull in a
Sucrase bug fix or when you want to use Sucrase in different ways in the same
project (e.g. in Jest and in Webpack). This PR removes the regular dependency in
favor of a peer dependency and updates the docs.