Implement the TypeScript satisfies operator
Fixes #763

In both the Babel and TS implementations, it's parsed exactly the same as the
`as` operator, so we can do the same here, and since all tokens are marked as
types, the transformer step naturally removes them.

I also got rid of an explicit `tt._as` assignment, which doesn't seem to be
necessary in this case.