Fix parsing of << within a type
Fixes #758

There was already some logic from #716 to handle cases like `f<<T>() => void>`,
where we're transitioning from a non-type context to a type context and need to
detect if `<<` is actually two open-type-parameter/argument tokens. There was a
simpler missing case, though, which is that `<<` is never allowed within a type,
so we should exit early with a simple `<` in a type context. There was already a
similar case for `>`, so this just uses the same logic for `<`. This fixes the
same issue in TS and Flow.