Add "preserve" mode for JSX
Fixes #780
Fixes #559

This PR adds support for preserving JSX as-is. Since the parser need to know if
JSX is enabled or not, preserving JSX is accomplished by keeping `"jsx"` as a
transform, but setting the `jsxRuntime` mode to `"preserve"`.

Most of the details just work: the imports and typescript transform are mostly
able to remove type syntax and transform imported names properly. The one
missing case was handling situations like `<div>`, which is not actually an
identifier access and thus should not be transformed by the imports transform.
This required adding a special case to the parser to remove the identifier role
in that case.

For now, the ts-node integration does not recognize this option because Node
wouldn't be able to recognize JSX anyway.