Fix dynamic import on plain CJS files (#789)
Fixes #778

Previously, `import()` transpiled to a promise-wrapped `require`, which worked
fine for ESM-to-CJS-transpiled modules, but when importing plain CJS modules,
the behavior was inconsistent with other implementations. In that case, we
needed to call `_interopRequireWildcard` to nest the module under a `default`
key.

I tested to confirm that this updated behavior is consistent with Node
ESM-to-CJS dynamic import, as well as Babel, TypeScript, and swc, so this change
will be considered a bug fix rather than a breaking change, even though it is
possible that existing use cases may have been relying on the old behavior.