Properly elide or preserve empty imports/exports (#810)
Fixes #801

Sucrase previously had a somewhat unified approach to import/export elision, but
this caused an issue with `import {}` statements, which should be removed in
TypeScript, but not in Flow or with neither transform enabled. Now, we plumb the
full configuration to both transforms and use the configuration to decide which
strategy is relevant.