Add CLI flags for all Sucrase options
Supersedes #670
Progress toward #792

This PR adds to the CLI a number of options that had been added. The CLI could
still use some work in terms of error reporting and behavior, but this should
hopefully get it closer to parity with other use cases.

The CLI previously had no tests, so this also adds an integration test suite
exercising each new and existing Sucrase option (though it's not yet at full
coverage).