Allow specifying options with sucrase/register and related tools
Fixes #792

This PR adds a new environment variable SUCRASE_OPTIONS that works with the
require hook and other tools that are based on the require hook (`sucrase-node`
and programmatically registering a hook). Its format is simply a JSON string
that is parsed and merged with the usual options. Hopefully this will be rare,
especially after a semver-major change to better defaults, but it's important
for some cases.